drop id setters; add 404 on delete when entity is missing

+++ core/src/main/java/org/sonatype/goodies/grafeas/internal/v1alpha1/NotesResource.java
@@ -163,6 +163,9 @@ public class NotesResource
 
     log.debug("Delete: {}/{}", project, name);
     NoteEntity entity = dao().read(project, name);
+    if (entity == null) {
+      throw new WebApplicationException(Status.NOT_FOUND);
+    }
     dao().delete(entity);
   }
